+++ apps/expo/features/climbs/WallScreen/index.tsx
@@ -1,7 +1,7 @@
 import wall from "@andescalada/api/schemas/wall";
 import { SoftDeleteSchema } from "@andescalada/db/zod";
 import useZodForm from "@andescalada/hooks/useZodForm";
-import { Box, LoadingModal, Pressable, Screen, Text } from "@andescalada/ui";
+import { LoadingModal, Screen } from "@andescalada/ui";
 import { trpc } from "@andescalada/utils/trpc";
 import Header from "@features/climbs/components/Header";
 import useHeaderOptionButton from "@features/climbs/components/HeaderOptionsButton/useHeaderOptions";
@@ -9,6 +9,7 @@ import {
   ClimbsNavigationRoutes,
   ClimbsNavigationScreenProps,
 } from "@features/climbs/Navigation/types";
+import OtherTopos from "@features/climbs/WallScreen/OtherTopos";
 import RoutesList from "@features/climbs/WallScreen/RoutesList";
 import TopoImage from "@features/climbs/WallScreen/TopoImage";
 import { MultiPitchManagerRoutes } from "@features/multiPitchManager/Navigation/types";
@@ -27,13 +28,6 @@ type Props = ClimbsNavigationScreenProps<ClimbsNavigationRoutes.Wall>;
 
 const WallScreen: FC<Props> = ({ route, navigation }) => {
   const { wallId, zoneId, sectorId, wallName, sectorKind } = route.params;
-
-  const otherToposQuery = trpc.topos.otherToposCount.useQuery({
-    wallId,
-    zoneId,
-  });
-
-  const otherToposCount = otherToposQuery.data ?? 0;
 
   const utils = trpc.useContext();
 
@@ -190,30 +184,7 @@ const WallScreen: FC<Props> = ({ route, navigation }) => {
         />
       </FormProvider>
       <TopoImage />
-      {otherToposCount > 0 && (
-        <Box paddingHorizontal="m" marginTop="m">
-          <Pressable
-            bg="grayscale.900"
-            borderRadius={8}
-            padding="s"
-            alignItems="center"
-            justifyContent="center"
-            gap="s"
-            flexDirection="row"
-            onPress={() =>
-              navigation.navigate(ClimbsNavigationRoutes.OtherTopos, {
-                wallId,
-                zoneId,
-              })
-            }
-          >
-            <Box bg="semantic.info" borderRadius={8} paddingHorizontal="s">
-              <Text variant="p3R">{otherToposCount}</Text>
-            </Box>
-            <Text variant="p2R">{`Ver otros topos`}</Text>
-          </Pressable>
-        </Box>
-      )}
+      <OtherTopos />
       <RoutesList />
     </Screen>
   );
